pyproject.toml
[project]
name = "prism-annotator"
version = "0.2.0"
description = "PRISM annotation of medical texts using LLMs and TANL parsing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "typer>=0.12",
]

[project.scripts]
prism = "prism_annotator.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/prism_annotator"]
